add strict tag level checking for iq xml
calibdb will validate every tag when parsing the iq xml file,
and if any tag attribute or tag tree is not the same as the
predefined pattern in calibdb, then the program will exit.

calidb: v1.0.0  checksum:635074

Old-Change-Id: Id7399ab381ef89feca337c7193b73ecdd7a674ef

+++ rkisp/ia-engine/include/common/cam_types.h
@@ -1828,7 +1828,7 @@ typedef struct CamCalibContrastAf_s {
   float 				  TrigThers;							  /**< AF trigger threshold */
   uint16_t				  TrigValue;							  /**< AF trigger Value */
   uint16_t				  TrigFrames;							  /**< AF trigger status must hold frames */
-  uint8_t				  TrigAntiFlash;						  /**< AF trigger anti one or some figures flash but not target figures*/
+  uint16_t				  TrigAntiFlash;						  /**< AF trigger anti one or some figures flash but not target figures*/
 
   float 				  FirstStableThers; 					  /**< first time AF stable threshold */
   uint16_t				  FirstStableValue; 					  /**< first time AF stable value */
